Sending chat und data with message id
