+++ include/roomMesseges.hpp
@@ -1,0 +1,18 @@
+//
+//  roomMesseges.hpp
+//  tunnler
+//
+
+#ifndef roomMesseges_h
+#define roomMesseges_h
+
+#include "MessageIdentifiers.h"
+
+enum RoomMessages
+{
+	ID_ROOM_CHAT = ID_USER_PACKET_ENUM + 1,
+	ID_ROOM_DATA = ID_USER_PACKET_ENUM + 2
+};
+
+
+#endif /* roomMesseges_h */
